+++ GTiff2Tiles.GUI/App.xaml.cs
@@ -1,8 +1,19 @@
-﻿namespace GTiff2Tiles.GUI
+﻿using Prism.Ioc;
+using GTiff2Tiles.GUI.Views;
+using System.Windows;
+
+namespace GTiff2Tiles.GUI
 {
-    /// <inheritdoc />
     /// <summary>
     /// Interaction logic for App.xaml
     /// </summary>
-    public partial class App { }
+    public partial class App
+    {
+        protected override Window CreateShell() => Container.Resolve<MainView>();
+
+        protected override void RegisterTypes(IContainerRegistry containerRegistry)
+        {
+
+        }
+    }
 }