+++ GTiff2Tiles.GUI/Views/MainView.xaml.cs
@@ -1,7 +1,7 @@
 ﻿namespace GTiff2Tiles.GUI.Views
 {
     /// <summary>
-    /// Interaction logic for Window1.xaml
+    /// Interaction logic for MainView.xaml
     /// </summary>
     public partial class MainView
     {